+++ src/backend/distributed/commands/extension.c
@@ -1,0 +1,117 @@
+/*-------------------------------------------------------------------------
+ *
+ * extension.c
+ *    Commands for creating and altering extensions.
+ *
+ * Copyright (c) 2018, Citus Data, Inc.
+ *
+ *-------------------------------------------------------------------------
+ */
+
+#include "postgres.h"
+
+#include "citus_version.h"
+#include "distributed/commands.h"
+#include "distributed/metadata_cache.h"
+#include "nodes/parsenodes.h"
+
+/* Local functions forward declarations for helper functions */
+static char * ExtractNewExtensionVersion(Node *parsetree);
+
+
+/*
+ * IsCitusExtensionStmt returns whether a given utility is a CREATE or ALTER
+ * EXTENSION statement which references the citus extension. This function
+ * returns false for all other inputs.
+ */
+bool
+IsCitusExtensionStmt(Node *parsetree)
+{
+	char *extensionName = "";
+
+	if (IsA(parsetree, CreateExtensionStmt))
+	{
+		extensionName = ((CreateExtensionStmt *) parsetree)->extname;
+	}
+	else if (IsA(parsetree, AlterExtensionStmt))
+	{
+		extensionName = ((AlterExtensionStmt *) parsetree)->extname;
+	}
+
+	return (strcmp(extensionName, "citus") == 0);
+}
+
+
+/*
+ * ErrorIfUnstableCreateOrAlterExtensionStmt compares CITUS_EXTENSIONVERSION
+ * and version given CREATE/ALTER EXTENSION statement will create/update to. If
+ * they are not same in major or minor version numbers, this function errors
+ * out. It ignores the schema version.
+ */
+void
+ErrorIfUnstableCreateOrAlterExtensionStmt(Node *parsetree)
+{
+	char *newExtensionVersion = ExtractNewExtensionVersion(parsetree);
+
+	if (newExtensionVersion != NULL)
+	{
+		/*  explicit version provided in CREATE or ALTER EXTENSION UPDATE; verify */
+		if (!MajorVersionsCompatible(newExtensionVersion, CITUS_EXTENSIONVERSION))
+		{
+			ereport(ERROR, (errmsg("specified version incompatible with loaded "
+								   "Citus library"),
+							errdetail("Loaded library requires %s, but %s was specified.",
+									  CITUS_MAJORVERSION, newExtensionVersion),
+							errhint("If a newer library is present, restart the database "
+									"and try the command again.")));
+		}
+	}
+	else
+	{
+		/*
+		 * No version was specified, so PostgreSQL will use the default_version
+		 * from the citus.control file.
+		 */
+		CheckAvailableVersion(ERROR);
+	}
+}
+
+
+/*
+ * ExtractNewExtensionVersion returns the new extension version specified by
+ * a CREATE or ALTER EXTENSION statement. Other inputs are not permitted. This
+ * function returns NULL for statements with no explicit version specified.
+ */
+static char *
+ExtractNewExtensionVersion(Node *parsetree)
+{
+	char *newVersion = NULL;
+	List *optionsList = NIL;
+	ListCell *optionsCell = NULL;
+
+	if (IsA(parsetree, CreateExtensionStmt))
+	{
+		optionsList = ((CreateExtensionStmt *) parsetree)->options;
+	}
+	else if (IsA(parsetree, AlterExtensionStmt))
+	{
+		optionsList = ((AlterExtensionStmt *) parsetree)->options;
+	}
+	else
+	{
+		/* input must be one of the two above types */
+		Assert(false);
+	}
+
+	foreach(optionsCell, optionsList)
+	{
+		DefElem *defElement = (DefElem *) lfirst(optionsCell);
+		if (strncmp(defElement->defname, "new_version", NAMEDATALEN) == 0)
+		{
+			newVersion = strVal(defElement->arg);
+			break;
+		}
+	}
+
+	return newVersion;
+}
