+++ src/backend/distributed/commands/grant.c
@@ -1,0 +1,19 @@
+/*-------------------------------------------------------------------------
+ *
+ * grant.c
+ *    Commands for granting access to distributed tables.
+ *
+ * Copyright (c) 2018, Citus Data, Inc.
+ *
+ *-------------------------------------------------------------------------
+ */
+
+#include "distributed/commands.h"
+
+
+/* placeholder for PlanGrantStmt */
+List *
+PlanGrantStmt(GrantStmt *grantStmt)
+{
+	return NIL;
+}
