+++ src/backend/distributed/commands/policy.c
@@ -12,8 +12,8 @@
 
 #include "catalog/namespace.h"
 #include "commands/policy.h"
+#include "distributed/commands.h"
 #include "distributed/metadata_cache.h"
-#include "distributed/policy.h"
 #include "utils/builtins.h"
