+++ src/backend/distributed/commands/transmit.c
@@ -15,6 +15,7 @@
 #include <sys/stat.h>
 #include <unistd.h>
 
+#include "commands/defrem.h"
 #include "distributed/relay_utility.h"
 #include "distributed/transmit.h"
 #include "distributed/worker_protocol.h"
@@ -329,4 +330,84 @@ ReceiveCopyData(StringInfo copyData)
 	}
 
 	return copyDone;
+}
+
+
+/* Is the passed in statement a transmit statement? */
+bool
+IsTransmitStmt(Node *parsetree)
+{
+	if (IsA(parsetree, CopyStmt))
+	{
+		CopyStmt *copyStatement = (CopyStmt *) parsetree;
+		ListCell *optionCell = NULL;
+
+		/* Extract options from the statement node tree */
+		foreach(optionCell, copyStatement->options)
+		{
+			DefElem *defel = (DefElem *) lfirst(optionCell);
+
+			if (strncmp(defel->defname, "format", NAMEDATALEN) == 0 &&
+				strncmp(defGetString(defel), "transmit", NAMEDATALEN) == 0)
+			{
+				return true;
+			}
+		}
+	}
+
+	return false;
+}
+
+
+/*
+ * VerifyTransmitStmt checks that the passed in command is a valid transmit
+ * statement. Raise ERROR if not.
+ *
+ * Note that only 'toplevel' options in the CopyStmt struct are checked, and
+ * that verification of the target files existance is not done here.
+ */
+void
+VerifyTransmitStmt(CopyStmt *copyStatement)
+{
+	char *fileName = NULL;
+
+	EnsureSuperUser();
+
+	/* do some minimal option verification */
+	if (copyStatement->relation == NULL ||
+		copyStatement->relation->relname == NULL)
+	{
+		ereport(ERROR, (errcode(ERRCODE_INVALID_PARAMETER_VALUE),
+						errmsg("FORMAT 'transmit' requires a target file")));
+	}
+
+	fileName = copyStatement->relation->relname;
+
+	if (is_absolute_path(fileName))
+	{
+		ereport(ERROR, (errcode(ERRCODE_INSUFFICIENT_PRIVILEGE),
+						(errmsg("absolute path not allowed"))));
+	}
+	else if (!path_is_relative_and_below_cwd(fileName))
+	{
+		ereport(ERROR,
+				(errcode(ERRCODE_INSUFFICIENT_PRIVILEGE),
+				 (errmsg("path must be in or below the current directory"))));
+	}
+
+	if (copyStatement->filename != NULL)
+	{
+		ereport(ERROR, (errcode(ERRCODE_INVALID_PARAMETER_VALUE),
+						errmsg("FORMAT 'transmit' only accepts STDIN/STDOUT"
+							   " as input/output")));
+	}
+
+	if (copyStatement->query != NULL ||
+		copyStatement->attlist != NULL ||
+		copyStatement->is_program)
+	{
+		ereport(ERROR, (errcode(ERRCODE_INVALID_PARAMETER_VALUE),
+						errmsg("FORMAT 'transmit' does not accept query, attribute list"
+							   " or PROGRAM parameters ")));
+	}
 }
