+++ src/backend/distributed/executor/multi_executor.c
@@ -16,9 +16,9 @@
 #include "catalog/dependency.h"
 #include "catalog/namespace.h"
 #include "distributed/citus_custom_scan.h"
+#include "distributed/commands/multi_copy.h"
 #include "distributed/insert_select_executor.h"
 #include "distributed/insert_select_planner.h"
-#include "distributed/multi_copy.h"
 #include "distributed/multi_executor.h"
 #include "distributed/multi_master_planner.h"
 #include "distributed/distributed_planner.h"
@@ -26,7 +26,6 @@
 #include "distributed/multi_router_planner.h"
 #include "distributed/multi_resowner.h"
 #include "distributed/multi_server_executor.h"
-#include "distributed/multi_utility.h"
 #include "distributed/resource_lock.h"
 #include "distributed/worker_protocol.h"
 #include "executor/execdebug.h"
