+++ src/backend/distributed/master/master_truncate.c
@@ -15,9 +15,10 @@
 
 #include "commands/tablecmds.h"
 #include "commands/trigger.h"
+#include "distributed/commands/utility_hook.h"
 #include "distributed/master_metadata_utility.h"
+#include "distributed/master_protocol.h"
 #include "distributed/multi_join_order.h"
-#include "distributed/multi_utility.h"
 #include "distributed/pg_dist_partition.h"
 #include "utils/builtins.h"
 #include "utils/lsyscache.h"
