+++ src/backend/distributed/metadata/metadata_sync.c
@@ -28,8 +28,8 @@
 #include "catalog/pg_namespace.h"
 #include "catalog/pg_type.h"
 #include "distributed/citus_ruleutils.h"
+#include "distributed/commands.h"
 #include "distributed/distribution_column.h"
-#include "distributed/foreign_constraint.h"
 #include "distributed/listutils.h"
 #include "distributed/master_metadata_utility.h"
 #include "distributed/master_protocol.h"
