+++ src/backend/distributed/planner/recursive_planning.c
@@ -55,10 +55,10 @@
 #include "catalog/pg_class.h"
 #include "distributed/citus_nodes.h"
 #include "distributed/citus_ruleutils.h"
+#include "distributed/commands/multi_copy.h"
 #include "distributed/distributed_planner.h"
 #include "distributed/errormessage.h"
 #include "distributed/metadata_cache.h"
-#include "distributed/multi_copy.h"
 #include "distributed/multi_logical_planner.h"
 #include "distributed/multi_router_planner.h"
 #include "distributed/multi_physical_planner.h"
