+++ src/backend/distributed/relay/relay_event_utility.c
@@ -29,8 +29,8 @@
 #include "catalog/namespace.h"
 #include "catalog/pg_class.h"
 #include "catalog/pg_constraint.h"
+#include "distributed/commands.h"
 #include "distributed/metadata_cache.h"
-#include "distributed/policy.h"
 #include "distributed/relay_utility.h"
 #include "distributed/version_compat.h"
 #include "lib/stringinfo.h"
