+++ src/include/distributed/commands/utility_hook.h
@@ -1,6 +1,6 @@
 /*-------------------------------------------------------------------------
  *
- * multi_utility.h
+ * utility_hook.h
  *	  Citus utility hook and related functionality.
  *
  * Copyright (c) 2012-2016, Citus Data, Inc.
@@ -10,10 +10,14 @@
 #ifndef MULTI_UTILITY_H
 #define MULTI_UTILITY_H
 
+#include "postgres.h"
+
+#include "distributed/version_compat.h"
+#include "utils/relcache.h"
 #include "tcop/utility.h"
 
 extern bool EnableDDLPropagation;
-extern bool EnableVersionChecks;
+
 
 /*
  * A DDLJob encapsulates the remote tasks and commands needed to process all or
@@ -30,9 +34,6 @@ typedef struct DDLJob
 	List *taskList;            /* worker DDL tasks to execute */
 } DDLJob;
 
-#if (PG_VERSION_NUM < 100000)
-struct QueryEnvironment; /* forward-declare to appease compiler */
-#endif
 
 extern void multi_ProcessUtility(PlannedStmt *pstmt, const char *queryString,
 								 ProcessUtilityContext context, ParamListInfo params,
@@ -44,14 +45,8 @@ extern void multi_ProcessUtility9x(Node *parsetree, const char *queryString,
 extern void CitusProcessUtility(Node *node, const char *queryString,
 								ProcessUtilityContext context, ParamListInfo params,
 								DestReceiver *dest, char *completionTag);
-extern List * PlanGrantStmt(GrantStmt *grantStmt);
-extern void ErrorIfUnsupportedConstraint(Relation relation, char distributionMethod,
-										 Var *distributionColumn, uint32 colocationId);
-
-extern Datum master_drop_all_shards(PG_FUNCTION_ARGS);
-extern Datum master_modify_multiple_shards(PG_FUNCTION_ARGS);
-
+extern void MarkInvalidateForeignKeyGraph(void);
+extern void InvalidateForeignKeyGraphForDDL(void);
 extern List * DDLTaskList(Oid relationId, const char *commandString);
-extern const char * RoleSpecString(RoleSpec *spec);
 
 #endif /* MULTI_UTILITY_H */
